requirements.txt
mysql-connector-python
pandas
numpy
matplotlib
seaborn
scikit-learn
flask
pickle